Extract screenshot methods to separate module
