+++ .rubocop_todo.yml
@@ -1,6 +1,6 @@
 # This configuration was generated by
 # `rubocop --auto-gen-config`
-# on 2016-11-07 18:55:47 +0300 using RuboCop version 0.45.0.
+# on 2016-11-07 19:04:38 +0300 using RuboCop version 0.45.0.
 # The point is for the user to remove these configuration records
 # one by one as the offenses are removed from the code base.
 # Note that changes in the inspected code, or installation of new
@@ -13,6 +13,7 @@ Lint/RescueException:
     - 'lib/onlyoffice_webdriver_wrapper/helpers/headless_helper/real_display_tools.rb'
     - 'lib/onlyoffice_webdriver_wrapper/webdriver.rb'
     - 'lib/onlyoffice_webdriver_wrapper/webdriver/webdriver_attributes_helper.rb'
+    - 'lib/onlyoffice_webdriver_wrapper/webdriver/webdriver_screenshot_helper.rb'
     - 'lib/onlyoffice_webdriver_wrapper/webdriver/webdriver_type_helper.rb'
 
 # Offense count: 11
@@ -22,7 +23,7 @@ Metrics/AbcSize:
 # Offense count: 1
 # Configuration parameters: CountComments.
 Metrics/ClassLength:
-  Max: 781
+  Max: 739
 
 # Offense count: 3
 Metrics/CyclomaticComplexity:
